services/data_ingestion/requirements.txt
pandas==2.1.2
psycopg2-binary==2.9.9